服务器导入protobuf 源码

+++ TcpClient/Assets/Scripts/Net/NetClient.cs
@@ -111,8 +111,8 @@ namespace Net
         {
             byte[] messageData = message.ToByteArray();
             int bodyLength = messageData.Length + NetPackage.MsgTypeLength;
-            byte[] bodyLengthData = BitConverter.GetBytes(bodyLength);
-            byte[] msgTypeData = BitConverter.GetBytes((int)msgType);
+            byte[] bodyLengthData = BitConverter.GetBytes((short)bodyLength);
+            byte[] msgTypeData = BitConverter.GetBytes((short)msgType);
             int length = bodyLength + NetPackage.HeadLength;//最终发送协议包长度
             byte[] sendBuffer = new byte[length];
             bodyLengthData.CopyTo(sendBuffer, 0);
